Build without CPP for ghc >= 8.0.0.
