+++ .gitignore
@@ -15,3 +15,5 @@ cabal.sandbox.config
 *.prof
 *.aux
 *.hp
+*.local
+*.swn
