+++ src-ghc-lib-8.10/GhcApi/GhcPlugins.hs
@@ -1,0 +1,3 @@
+module GhcApi.GhcPlugins (module GhcPlugins) where
+
+import GhcPlugins hiding (TcPlugin, mkSubst)
