+++ src-ghc-lib-8.2/GhcApi/GhcPlugins.hs
@@ -1,0 +1,5 @@
+module GhcApi.GhcPlugins (module GhcPlugins, TcTyVar, EvTerm(..)) where
+
+import GhcPlugins hiding (mkSubst)
+import TcType (TcTyVar)
+import TcEvidence (EvTerm(..))
