+++ src-ghc-lib-8.6/GhcApi/GhcPlugins.hs
@@ -1,0 +1,3 @@
+module GhcApi.GhcPlugins (module GhcPlugins) where
+
+import GhcPlugins hiding (mkSubst)
