+++ src-ghc-lib/GhcApi/Constraint.hs
@@ -1,0 +1,12 @@
+module GhcApi.Constraint
+  ( Ct(..)
+  , CtEvidence(..)
+  , CtLoc
+  , ctLoc
+  , ctEvId
+  , mkNonCanonical
+  )
+where
+
+import Constraint
+  (Ct (..), CtEvidence (..), CtLoc, ctLoc, ctEvId, mkNonCanonical)
