+++ src-ghc-lib/GhcApi/Predicate.hs
@@ -1,0 +1,3 @@
+module GhcApi.Predicate (mkPrimEqPred) where
+
+import Predicate (mkPrimEqPred)
